requirements.txt
streamlit
pdfminer.six
docx2txt
langchain
chromadb
sentence-transformers
langchain-community
